基本完成后端

+++ web-menagement/src/main/java/com/example/service/impl/DeptServiceimpl.java
@@ -1,5 +1,6 @@
 package com.example.service.impl;
 
+import com.example.exception.BusinessException;
 import com.example.mapper.DeptMapper;
 import com.example.pojo.Dept;
 import com.example.service.DeptService;
@@ -19,8 +20,13 @@ public class DeptServiceimpl implements DeptService {
         return  deptMapper.findAll();
     }
 
+    //如果部门有员工不能删除
     @Override
     public void deleteById(Integer id) {
+        if (deptMapper.findEmpByDeptId(id) > 0) {
+            //提示前端
+            throw new BusinessException("该部门有员工，不能删除");
+            }
         deptMapper.deleteById(id);
     }
